+++ include/stp/Sat/Riss.h
@@ -72,7 +72,7 @@ public:
 
   void setVerbosity(int v) override;
 
-  unsigned long nVars() const override;
+  uint32_t nVars() const override;
 
   void printStats() const override;
